+++ tests/test_control.py
@@ -19,7 +19,7 @@ from context import Context
 from tree import *
 from eval import rootContext, moduleContext
 
-from strformat import *
+from bases.strformat import *
 
 
 from tests.utils import *
