+++ tests/test_format.py
@@ -17,7 +17,7 @@ from context import Context
 from tree import *
 from eval import rootContext
 
-from strformat import *
+from bases.strformat import *
 
 import pdb
