+++ tests/test_module.py
@@ -11,7 +11,7 @@ from cases.utils import *
 from nodes.structs import *
 from tree import *
 from eval import rootContext, moduleContext
-from strformat import *
+from bases.strformat import *
 from loader import *
 
 from tests.utils import *
