+++ tests/test_vals.py
@@ -8,7 +8,7 @@ from parser import splitLine, splitLexems, charType, splitOper, elemStream
 from vars import *
 from vals import numLex
 from context import Context
-from strformat import *
+from bases.strformat import *
 from nodes.structs import *
 from tree import *
 from eval import rootContext, moduleContext
